+++ include/heyoka/math/time.hpp
@@ -11,6 +11,7 @@
 
 #include <cstdint>
 #include <ostream>
+#include <string>
 #include <vector>
 
 #include <heyoka/config.hpp>
@@ -32,6 +33,8 @@ public:
     time_impl();
 
     void to_stream(std::ostream &) const;
+
+    expression diff(const std::string &) const;
 
     llvm::Value *taylor_diff_dbl(llvm_state &, const std::vector<std::uint32_t> &, const std::vector<llvm::Value *> &,
                                  llvm::Value *, llvm::Value *, std::uint32_t, std::uint32_t, std::uint32_t,